+++ FOSSDB_web/apps/fossdb/models.py
@@ -6,15 +6,6 @@ from django.db import models
 User = settings.AUTH_USER_MODEL
 
 
-class Tag(models.Model):
-    name = models.CharField(max_length=100)
-    description = models.TextField(blank=True, default="")
-    icon = models.ImageField(upload_to="types/icons/", null=True, blank=True)
-
-    def __str__(self):
-        return self.name
-
-
 class Project(models.Model):
     uuid = models.UUIDField(primary_key=True, default=uuid.uuid4, editable=False)
     author = models.ForeignKey(User, on_delete=models.CASCADE)
@@ -23,7 +14,7 @@ class Project(models.Model):
     licenses = models.ManyToManyField("License")
     programming_languages = models.ManyToManyField("ProgrammingLanguage", through="ProjectProgrammingLanguage", related_name="projects")
     hosting_platform = models.ManyToManyField("HostingPlatform", through="ProjectHostingPlatform", related_name="projects")
-    tag = models.ManyToManyField(Tag)
+    tag = models.ManyToManyField("Tag")
     date_created = models.DateTimeField(auto_now_add=True)
 
     def save(self, *args, **kwargs):
